+++ src/Feature/Sms/Bag/ScheduleSmssBag.php
@@ -11,7 +11,6 @@ use DateTimeInterface;
  * @property string $from
  * @property string $message
  * @property string $template
- * @property string $encoding
  * @property array $idx
  * @property bool $checkIdx
  * @property string $partnerId
@@ -34,6 +33,9 @@ class ScheduleSmssBag
 
     /** @var DateTimeInterface */
     public $date;
+
+    /** @var string */
+    public $encoding = 'utf-8';
 
     public static function withMessage(DateTimeInterface $scheduleAt, array $receivers, string $message): self
     {