+++ src/Feature/Sms/Bag/SendSmsBag.php
@@ -10,7 +10,6 @@ use DateTimeInterface;
  * @property string $from
  * @property string $message
  * @property string $template
- * @property string $encoding
  * @property array $idx
  * @property bool $checkIdx
  * @property string $partnerId
@@ -30,6 +29,9 @@ class SendSmsBag
 {
     /** @var string */
     public $to;
+
+    /** @var string */
+    public $encoding = 'utf-8';
 
     public static function withMessage(string $receiver, string $message): self
     {