+++ src/Feature/Sms/Bag/SendSmsToGroupBag.php
@@ -10,7 +10,6 @@ use DateTimeInterface;
  * @property string $from
  * @property string $message
  * @property string $template
- * @property string $encoding
  * @property array $idx
  * @property bool $checkIdx
  * @property string $partnerId
@@ -25,6 +24,9 @@ class SendSmsToGroupBag
 {
     /** @var string */
     public $group;
+
+    /** @var string */
+    public $encoding = 'utf-8';
 
     public static function withMessage(string $group, string $message): self
     {